+++ commands.go
@@ -21,7 +21,7 @@ import (
 	"unicode"
 )
 
-const VERSION = "0.3.1"
+const VERSION = "0.3.2"
 
 var (
 	GIT_COMMIT string